Refactor "pypw.generate" function signature
Refactor the pypw.generate function to accept a string argument,
instead of a dictionary, to represent the chosen word list. Also
refactor the branch logic to deal with this change. Update the
docstring for the generate function. Add return statement to the
generate function so that it returns a List containing the words
which make up the generated password.