Add Sphinx files
Add conf.py, index.rst, and Makefile for Sphinx documentation
system. These files are located under the "doc" folder which lives
in the top level of the project hierarchy. The way Sphinx is set
up for this project, there are three subfolders in the doc folder,
doc/_build/, doc/_static/, and doc/_templates/. These are not
tracked in git currently because I do not think they are
necessary but I will test to see if it builds without them and
will add them if it turns out they are needed.